ARTEMIS-145 moving compact command under tools
https://issues.apache.org/jira/browse/ARTEMIS-145
